+++ include/mruby_integration/models/monitor.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "mruby.h"
+
+void
+append_models_Monitor(mrb_state*);