+++ include/mruby_integration/models/window.hpp
@@ -2,5 +2,7 @@
 
 #include "mruby.h"
 
+extern struct RClass* Window_class;
+
 void
 append_models_Window(mrb_state*);