+++ src/ruby/models/monitor.rb
@@ -1,0 +1,4 @@
+# This class is used for retrieving information about the monitors attached to
+# the computer.
+module Monitor
+end